fixing a bug

+++ tests_and_random_things/pensamento_computacional/atv3.py
@@ -3,7 +3,7 @@
 temperatura = input().split()
 
 if temperatura[1] == "C":
-    valor = int(temperatura[0])
+    valor = float(temperatura[0])
     fahrenheit = (valor * 1.8) + 32
     kelvin = valor + 273.15
     print(f"Temperatura em Celsius: {valor:.2f} °C")
@@ -11,7 +11,7 @@ if temperatura[1] == "C":
     print(f"Temperatura em Kelvin: {kelvin:.2f} K")
 
 elif temperatura[1] == "F":
-    valor = int(temperatura[0])
+    valor = float(temperatura[0])
     celsius = (valor - 32) / 1.8
     kelvin = ((valor - 32) / 1.8) + 273.15
     print(f"Temperatura em Celsius: {celsius:.2f} °C")
@@ -19,7 +19,7 @@ elif temperatura[1] == "F":
     print(f"Temperatura em Kelvin: {kelvin:.2f} K")
 
 elif temperatura[1] == "K":
-    valor = int(temperatura[0])
+    valor = float(temperatura[0])
     celsius = valor - 273.15
     fahrenheit = ((valor - 273.15) * 1.8) + 32 
     print(f"Temperatura em Celsius: {celsius:.2f} °C")
